# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Cambase.io is a project by Evercam.io in order to make it easier for\
    \ software developers to have a reliable, up to date source of model hardware\
    \ information available via a public API."
  version: "1.0"
  title: "Cambase.io"
  x-logo:
    url: "https://api.apis.guru/v2/cache/logo/https_twitter.com_evrcm_profile_image.jpeg"
  x-origin:
  - converter:
      url: "https://github.com/lucybot/api-spec-converter"
      version: "2.6.2"
    format: "swagger"
    url: "http://api.cambase.io/api-docs/api-docs.json"
    version: "1.2"
  x-preferred: true
  x-providerName: "cambase.io"
host: "api.cambase.io"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/v1/models/{id}.json:
    get:
      tags:
      - "models"
      summary: "Fetches a single Model"
      operationId: "Api::V1::Models#show"
      parameters:
      - name: "id"
        in: "path"
        description: "Model ID"
        required: true
        type: "integer"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Models"
    put:
      tags:
      - "models"
      summary: "Updates an existing Model"
      operationId: "apiV1ModelsId_jsonPUT"
      parameters:
      - name: "id"
        in: "path"
        description: "Model ID"
        required: true
        type: "string"
      - name: "vendor_id"
        in: "formData"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "model[model]"
        in: "formData"
        description: "Model"
        required: false
        type: "string"
      - name: "model[shape]"
        in: "formData"
        description: "Shape"
        required: false
        type: "string"
      - name: "model[resolution]"
        in: "formData"
        description: "Resolution"
        required: false
        type: "string"
      - name: "model[onvif]"
        in: "formData"
        description: "ONVIF"
        required: false
        type: "string"
      - name: "model[psia]"
        in: "formData"
        description: "PSIA"
        required: false
        type: "string"
      - name: "model[ptz]"
        in: "formData"
        description: "PTZ"
        required: false
        type: "string"
      - name: "model[infrared]"
        in: "formData"
        description: "Infrared"
        required: false
        type: "string"
      - name: "model[varifocal]"
        in: "formData"
        description: "Varifocal"
        required: false
        type: "string"
      - name: "model[sd_card]"
        in: "formData"
        description: "SD Card"
        required: false
        type: "string"
      - name: "model[upnp]"
        in: "formData"
        description: "UPnP"
        required: false
        type: "string"
      - name: "model[audio_in]"
        in: "formData"
        description: "Audio In"
        required: false
        type: "string"
      - name: "model[audio_out]"
        in: "formData"
        description: "Audio Out"
        required: false
        type: "string"
      - name: "model[default_username]"
        in: "formData"
        description: "Default Username"
        required: false
        type: "string"
      - name: "model[default_password]"
        in: "formData"
        description: "Default Password"
        required: false
        type: "string"
      - name: "model[jpeg_url]"
        in: "formData"
        description: "JPEG URL"
        required: false
        type: "string"
      - name: "model[h264_url]"
        in: "formData"
        description: "H264 URL"
        required: false
        type: "string"
      - name: "model[mjpeg_url]"
        in: "formData"
        description: "MJPEG URL"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Models"
    patch:
      tags:
      - "models"
      summary: "Updates an existing Model"
      operationId: "apiV1ModelsId_jsonPATCH"
      parameters:
      - name: "id"
        in: "path"
        description: "Model ID"
        required: true
        type: "string"
      - name: "vendor_id"
        in: "formData"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "model[model]"
        in: "formData"
        description: "Model"
        required: false
        type: "string"
      - name: "model[shape]"
        in: "formData"
        description: "Shape"
        required: false
        type: "string"
      - name: "model[resolution]"
        in: "formData"
        description: "Resolution"
        required: false
        type: "string"
      - name: "model[onvif]"
        in: "formData"
        description: "ONVIF"
        required: false
        type: "string"
      - name: "model[psia]"
        in: "formData"
        description: "PSIA"
        required: false
        type: "string"
      - name: "model[ptz]"
        in: "formData"
        description: "PTZ"
        required: false
        type: "string"
      - name: "model[infrared]"
        in: "formData"
        description: "Infrared"
        required: false
        type: "string"
      - name: "model[varifocal]"
        in: "formData"
        description: "Varifocal"
        required: false
        type: "string"
      - name: "model[sd_card]"
        in: "formData"
        description: "SD Card"
        required: false
        type: "string"
      - name: "model[upnp]"
        in: "formData"
        description: "UPnP"
        required: false
        type: "string"
      - name: "model[audio_in]"
        in: "formData"
        description: "Audio In"
        required: false
        type: "string"
      - name: "model[audio_out]"
        in: "formData"
        description: "Audio Out"
        required: false
        type: "string"
      - name: "model[default_username]"
        in: "formData"
        description: "Default Username"
        required: false
        type: "string"
      - name: "model[default_password]"
        in: "formData"
        description: "Default Password"
        required: false
        type: "string"
      - name: "model[jpeg_url]"
        in: "formData"
        description: "JPEG URL"
        required: false
        type: "string"
      - name: "model[h264_url]"
        in: "formData"
        description: "H264 URL"
        required: false
        type: "string"
      - name: "model[mjpeg_url]"
        in: "formData"
        description: "MJPEG URL"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Models"
...